cleanup init
