--- fonts/src/index.ts
@@ -1,56 +1,0 @@
-export const fontConfig = {
-  web: {
-    regular: {
-      fontFamily: 'Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif',
-      fontWeight: '400' as '400',
-    },
-    medium: {
-      fontFamily: 'Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif',
-      fontWeight: '500' as '500',
-    },
-    light: {
-      fontFamily: 'Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif',
-      fontWeight: '300' as '300',
-    },
-    thin: {
-      fontFamily: 'Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif',
-      fontWeight: '100' as '100',
-    },
-  },
-  ios: {
-    regular: {
-      fontFamily: 'System, Roboto',
-      fontWeight: '400' as '400',
-    },
-    medium: {
-      fontFamily: 'System, Roboto',
-      fontWeight: '500' as '500',
-    },
-    light: {
-      fontFamily: 'System, Roboto',
-      fontWeight: '300' as '300',
-    },
-    thin: {
-      fontFamily: 'System, Roboto',
-      fontWeight: '100' as '100',
-    },
-  },
-  default: {
-    regular: {
-      fontFamily: 'sans-serif',
-      fontWeight: '400' as '400',
-    },
-    medium: {
-      fontFamily: 'sans-serif',
-      fontWeight: '500' as '500',
-    },
-    light: {
-      fontFamily: 'sans-serif',
-      fontWeight: '300' as '300',
-    },
-    thin: {
-      fontFamily: 'sans-serif',
-      fontWeight: '100' as '100',
-    },
-  },
-};
