--- fonts/stories/index.stories.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import 'react-native';
-import { FontShowcase } from '~/showcases';
-
-export const fonts = () => <FontShowcase />;
