+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@divriots/starter-reactnative",
+  "name": "@divriots/starter-reactnative-paper",
   "version": "0.0.0",
   "description": "",
   "license": "ISC",
