+++ typography/stories/index.stories.tsx
@@ -1,17 +1,5 @@
-import React from 'react';;
-import { View } from 'react-native';
-import { Text, Title, Subheading, Paragraph, Headline, Caption } from '../index';
-import { storyLayout } from '~/layout';
+import React from 'react';
+import 'react-native';
+import { FontShowcase } from '~/showcases';
 
-export default storyLayout;
-
-export const typography = () => (
-  <View style={{ diplay: 'flex', flexDirection: 'column' }}>
-    <Text>Text</Text>
-    <Title>Title</Title>
-    <Subheading>Subheading</Subheading>
-    <Paragraph>Paragraph</Paragraph>
-    <Headline>Headline</Headline>
-    <Caption>Caption</Caption>
-  </View>
-);
+export const fonts = () => <FontShowcase />;
